GH-18 Application methods lifted to Opts

+++ enki/src/main/scala/enki/tests/EnkiSuite.scala
@@ -12,7 +12,7 @@ trait EnkiSuite extends Defaults with ImplicitConversions {
       .builder()
       .appName("enki-test")
       .master(s"local[${Runtime.getRuntime.availableProcessors}]")
-      .config("spark.sql.shuffle.partitions", Runtime.getRuntime.availableProcessors)
+      .config("spark.sql.shuffle.partitions", Runtime.getRuntime.availableProcessors.toLong)
       .config("spark.sql.warehouse.dir", Files.createTempDirectory("spark-warehouse").toUri.toString)
       .getOrCreate()
   }
